opa-react: rename opa-provider.tsx -> authz-provider.tsx
