+++ packages/opa-react/src/index.ts
@@ -1,6 +1,6 @@
 export { default as Authz } from "./authz.js";
 export type * from "./authz.js";
-export { default as AuthzProvider } from "./opa-provider.js";
-export type * from "./opa-provider.js";
+export { default as AuthzProvider } from "./authz-provider.js";
+export type * from "./authz-provider.js";
 export { default as useAuthz } from "./use-authz.js";
 export type * from "./use-authz.js";
